externalize join lobby room logic
