+++ app/public/src/pages/lobby.tsx
@@ -1,53 +1,11 @@
-import { type NonFunctionPropNames } from "@colyseus/schema/lib/types/HelperTypes"
-import { Client, Room, RoomAvailable } from "colyseus.js"
+import { RoomAvailable } from "colyseus.js"
 import firebase from "firebase/compat/app"
 import React, { useCallback, useEffect, useRef, useState } from "react"
 import { useTranslation } from "react-i18next"
 import { useNavigate } from "react-router-dom"
-import {
-  TournamentBracketSchema,
-  TournamentPlayerSchema,
-  TournamentSchema
-} from "../../../models/colyseus-models/tournament"
-import { IUserMetadata } from "../../../models/mongo-models/user-metadata"
-import {
-  ICustomLobbyState,
-  ISuggestionUser,
-  PkmWithConfig,
-  Transfer
-} from "../../../types"
-import { CloseCodes, CloseCodesMessages } from "../../../types/enum/CloseCodes"
-import { logger } from "../../../utils/logger"
 import { useAppDispatch, useAppSelector } from "../hooks"
-import store from "../stores"
-import {
-  addRoom,
-  addTournament,
-  addTournamentBracket,
-  changeTournament,
-  changeTournamentBracket,
-  changeTournamentPlayer,
-  leaveLobby,
-  pushBotLog,
-  pushMessage,
-  removeMessage,
-  removeRoom,
-  removeTournament,
-  removeTournamentBracket,
-  setBoosterContent,
-  setCcu,
-  setNextSpecialGame,
-  setSearchedUser,
-  setSuggestions,
-  updateTournament
-} from "../stores/LobbyStore"
-import {
-  joinLobby,
-  logIn,
-  logOut,
-  setErrorAlertMessage,
-  setProfile
-} from "../stores/NetworkStore"
+import { leaveLobby } from "../stores/LobbyStore"
+import { logOut, setErrorAlertMessage } from "../stores/NetworkStore"
 import { Announcements } from "./component/announcements/announcements"
 import AvailableRoomMenu from "./component/available-room-menu/available-room-menu"
 import { GameRoomsMenu } from "./component/available-room-menu/game-rooms-menu"
@@ -56,7 +14,7 @@ import { MainSidebar } from "./component/main-sidebar/main-sidebar"
 import { Modal } from "./component/modal/modal"
 import { cc } from "./utils/jsx"
 import { LocalStoreKeys, localStore } from "./utils/store"
-import { FIREBASE_CONFIG } from "./utils/utils"
+import { joinLobbyRoom } from "../game/lobby-logic"
 import "./lobby.css"
 
 export default function Lobby() {
@@ -64,7 +22,6 @@ export default function Lobby() {
   const navigate = useNavigate()
   const lobby = useAppSelector((state) => state.network.lobby)
 
-  const [profileRetries, setProfileRetries] = useState<number>(0)
   const [gameToReconnect, setGameToReconnect] = useState<string | null>(
     localStore.get(LocalStoreKeys.RECONNECTION_GAME)
   )
@@ -78,26 +35,13 @@ export default function Lobby() {
 
   const { t } = useTranslation()
 
-  const onLeave = (code: number) => {
-    logger.info(`left lobby with code ${code}`)
-    const errorMessage = CloseCodesMessages[code]
-    if (errorMessage) {
-      dispatch(setErrorAlertMessage(t(`errors.${errorMessage}`)))
-    }
-    navigate("/")
-  }
-
+  const lobbyJoined = useRef<boolean>(false)
   useEffect(() => {
-    const client = store.getState().network.client
-    joinLobbyRoom(dispatch, client, onLeave, setProfileRetries).catch((err) => {
-      logger.error(err)
-      const errorMessage = CloseCodesMessages[err] ?? "UNKNOWN_ERROR"
-      if (errorMessage) {
-        dispatch(setErrorAlertMessage(t(`errors.${errorMessage}`, { error: err })))
-      }
-      navigate("/")
-    })
-  }, [dispatch, profileRetries])
+    if (!lobbyJoined.current) {
+      joinLobbyRoom(dispatch, navigate)
+      lobbyJoined.current = true
+    }
+  }, [lobbyJoined])
 
   const signOut = useCallback(async () => {
     if (lobby?.connection.isOpen) {
@@ -223,213 +167,4 @@ function MainLobby() {
       </section>
     </div>
   )
-}
-
-export async function joinLobbyRoom(
-  dispatch,
-  client: Client,
-  onLeave: (code: number) => void,
-  setProfileRetries?: (value: React.SetStateAction<number>) => void
-): Promise<Room<ICustomLobbyState>> {
-  if (!firebase.apps.length) {
-    firebase.initializeApp(FIREBASE_CONFIG)
-  }
-  return new Promise((resolve, reject) => {
-    firebase.auth().onAuthStateChanged(async (user) => {
-      if (user) {
-        dispatch(logIn(user))
-        try {
-          const token = await user.getIdToken()
-          let lobby: Room<ICustomLobbyState> | undefined = store.getState().network.lobby
-          if (lobby?.connection.isOpen) {
-            localStore.set(LocalStoreKeys.RECONNECTION_LOBBY, lobby.reconnectionToken, 30)
-            lobby.onLeave.clear()
-            await lobby.leave(false)
-            lobby = undefined
-          }
-          const reconnectToken: string = localStore.get(LocalStoreKeys.RECONNECTION_LOBBY)
-          if (reconnectToken) {
-            try {
-              lobby = await client.reconnect(reconnectToken)
-            } catch (error) {
-              localStore.delete(LocalStoreKeys.RECONNECTION_LOBBY)
-            }
-          }
-          const room: Room<ICustomLobbyState> = lobby ?? await client.join("lobby", {
-            idToken: token
-          })
-          // store key for 5 minutes, but can still only rejoin within 30 seconds of leaving
-          localStore.set(LocalStoreKeys.RECONNECTION_LOBBY, room.reconnectionToken, 60 * 5)
-
-          room.onLeave(onLeave)
-
-          room.state.messages.onAdd((m) => {
-            dispatch(pushMessage(m))
-          })
-          room.state.messages.onRemove((m) => {
-            dispatch(removeMessage(m))
-          })
-
-          room.state.listen("ccu", (value) => {
-            dispatch(setCcu(value))
-          })
-
-          room.state.tournaments.onAdd((tournament) => {
-            dispatch(addTournament(tournament))
-            const fields: NonFunctionPropNames<TournamentSchema>[] = [
-              "id",
-              "name",
-              "startDate"
-            ]
-
-            fields.forEach((field) => {
-              tournament.listen(field, (value) => {
-                dispatch(
-                  changeTournament({
-                    tournamentId: tournament.id,
-                    field: field,
-                    value: value
-                  })
-                )
-              })
-            })
-
-            tournament.players.onAdd((player, userId) => {
-              dispatch(updateTournament()) // TOFIX: force redux reactivity
-              const fields: NonFunctionPropNames<TournamentPlayerSchema>[] = [
-                "eliminated"
-              ]
-              fields.forEach((field) => {
-                player.listen(field, (value) => {
-                  dispatch(
-                    changeTournamentPlayer({
-                      tournamentId: tournament.id,
-                      playerId: userId,
-                      field: field,
-                      value: value
-                    })
-                  )
-                })
-              })
-            })
-
-            tournament.players.onRemove((player, userId) => {
-              dispatch(updateTournament()) // TOFIX: force redux reactivity
-            })
-
-            tournament.brackets.onAdd((bracket, bracketId) => {
-              dispatch(
-                addTournamentBracket({
-                  tournamendId: tournament.id,
-                  bracketId,
-                  bracket
-                })
-              )
-
-              const fields: NonFunctionPropNames<TournamentBracketSchema>[] = [
-                "name",
-                "finished"
-              ]
-              fields.forEach((field) => {
-                bracket.listen(field, (value) => {
-                  dispatch(
-                    changeTournamentBracket({
-                      tournamentId: tournament.id,
-                      bracketId,
-                      field,
-                      value
-                    })
-                  )
-                })
-              })
-
-              bracket.playersId.onChange(() => {
-                dispatch(
-                  changeTournamentBracket({
-                    tournamentId: tournament.id,
-                    bracketId,
-                    field: "playersId",
-                    value: bracket.playersId
-                  })
-                )
-              })
-            })
-
-            tournament.brackets.onRemove((bracket, bracketId) => {
-              dispatch(
-                removeTournamentBracket({
-                  tournamendId: tournament.id,
-                  bracketId
-                })
-              )
-            })
-          })
-
-          room.state.tournaments.onRemove((tournament) => {
-            dispatch(removeTournament(tournament))
-          })
-
-          room.state.listen("nextSpecialGame", (specialGame) => {
-            dispatch(setNextSpecialGame(specialGame))
-          })
-
-          room.onMessage(Transfer.BANNED, (message) => {
-            alert(message)
-          })
-
-          room.onMessage(Transfer.BOT_DATABASE_LOG, (message) => {
-            dispatch(pushBotLog(message))
-          })
-
-          room.onMessage(Transfer.ROOMS, (rooms: RoomAvailable[]) => {
-            rooms.forEach((room) => dispatch(addRoom(room)))
-          })
-
-          room.onMessage(Transfer.ADD_ROOM, ([, room]) => {
-            if (room.name === "preparation" || room.name === "game") {
-              dispatch(addRoom(room))
-            }
-          })
-
-          room.onMessage(Transfer.REMOVE_ROOM, (roomId: string) =>
-            dispatch(removeRoom(roomId))
-          )
-
-          room.onMessage(Transfer.USER_PROFILE, (user: IUserMetadata) => {
-            if (user) {
-              dispatch(setProfile(user))
-            } else if (setProfileRetries) {
-              setProfileRetries(x => x + 1)
-            }
-          })
-
-          room.onMessage(Transfer.USER, (user: IUserMetadata) =>
-            dispatch(setSearchedUser(user))
-          )
-
-          room.onMessage(
-            Transfer.BOOSTER_CONTENT,
-            (boosterContent: PkmWithConfig[]) => {
-              dispatch(setBoosterContent(boosterContent))
-            }
-          )
-
-          room.onMessage(
-            Transfer.SUGGESTIONS,
-            (suggestions: ISuggestionUser[]) => {
-              dispatch(setSuggestions(suggestions))
-            }
-          )
-
-          dispatch(joinLobby(room))
-
-          resolve(room)
-        } catch (error) {
-          reject(error)
-        }
-      } else {
-        reject(CloseCodes.USER_NOT_AUTHENTICATED)
-      }
-    })
-  })
 }
